Extract out a CassandraQueryRunner, and move the run stuff there
